split between Object and Package
